refactor(runtime): centralize filesystem-containment guards
The path-safety guards existed as private copies across the runtime skill
reader and the desktop main process. Move them verbatim into one shared
leaf module, `packages/runtime/src/path-containment.ts`, re-exported from
`@maka/runtime`, so a future fix to a containment check lands in one place.

- `isContainedPath` / `isSafeSkillId` were byte-identical in
  `runtime/skills.ts` and `managed-skill-sources.ts`; both now import the
  shared home (desktop main already depends on `@maka/runtime`).
- `explore-agent-tool.ts`'s `isInside` / `toRelative` are a distinct,
  separator-aware family (verified not byte-identical to `isContainedPath`);
  moved verbatim and imported, with the semantic divergence documented in
  the shared module so it stays a deliberate, behavior-changing decision.

Behavior-neutral: guards moved verbatim, no logic changed. Public
`@maka/runtime` exports stay stable so the later `skills.ts` split composes.

Refs #1084
